+++ src/Infocom.php
@@ -1425,15 +1425,6 @@ JS;
     }
 
     /**
-     * @deprecated
-     */
-    public static function addPluginInfos(CommonDBTM $item)
-    {
-        Toolbox::deprecated('Infocom::addPluginInfos() is deprecated. Use Plugin::doHookFunction(Hooks::INFOCOM, $item) directly instead.');
-        Plugin::doHookFunction(Hooks::INFOCOM, $item);
-    }
-
-    /**
      * @param $itemtype
      **/
     public static function rawSearchOptionsToAdd($itemtype = null)