+++ CHANGELOG.md
@@ -309,7 +309,6 @@ The present file will list all changes made to the project; according to the
 - `Html::jsSetDropdownValue()`
 - `Html::progressBar()`
 - `HookManager::enableCSRF()`
-- `Infocom::addPluginInfos()`
 - `ITILFollowup::ADDMYTICKET` constant. Use `ITILFollowup::ADDMY`.
 - `ITILFollowup::ADDGROUPTICKET` constant. Use `ITILFollowup::ADD_AS_GROUP`.
 - `ITILFollowup::ADDALLTICKET` constant. Use `ITILFollowup::ADDALLITEM`.
@@ -483,6 +482,7 @@ The present file will list all changes made to the project; according to the
 - `Html::openArrowMassives()`
 - `Html::showTimeField()`
 - `Impact::buildNetwork()`
+- `Infocom::addPluginInfos()`
 - `Infocom::showDebug()`
 - `IPNetwork::recreateTree()`
 - `IPNetwork::title()`